mobx: Move stores into their own folders
