+++ public/app/containers/AlertRuleList/AlertRuleList.jest.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import moment from 'moment';
 import { AlertRuleList } from './AlertRuleList';
-import { RootStore } from 'app/stores/RootStore';
+import { RootStore } from 'app/stores/RootStore/RootStore';
 import { backendSrv, createNavTree } from 'test/mocks/common';
 import { mount } from 'enzyme';
 import toJson from 'enzyme-to-json';
