+++ public/app/stores/ViewStore/ViewStore.ts
